embedding_pipeline/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
cohere==4.9.0
qdrant-client==1.8.0
python-dotenv==1.0.0
tiktoken==0.5.2
lxml==4.9.3